refactor (utils): centralizes auto-commit decorator
